+++ laktory/spark.py
@@ -1,4 +1,9 @@
-from pyspark.sql.utils import AnalysisException
+try:
+    from pyspark.sql import DataFrame
+    from pyspark.sql.utils import AnalysisException
+except ModuleNotFoundError:
+    class DataFrame:
+        pass
 
 
 def df_has_column(sdf, col):