+++ CHANGELOG.md
@@ -5,14 +5,14 @@
 - Keyvault support for Settings
 
 **Updated**
-- 
+- Refactored dlt module to support DBR < 13 and clusters in shared access mode
 
 **Fixed**
 - model_validate_yaml() for Pipeline model
 - table data insert when no data is available
 
 **Breaking Change**
--
+- Removed spark from required dependencies
 
 ## [0.0.3] - 2023-09-25
 **Added**